backend/requirements.txt
asgiref==3.4.1
certifi==2021.10.8
cloudinary==1.25.0
Django==3.2.7
django-cors-headers==3.8.0
django-filter==21.1
djangorestframework==3.12.4
Pillow==8.4.0
pytz==2021.1
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.7
